+++ src/RcppExports.cpp
@@ -2103,6 +2103,25 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// saemSeedLayoutTest_
+Rcpp::NumericVector saemSeedLayoutTest_(Rcpp::IntegerVector nu, int nu1B, int nphi1, int nphi0, int nMix, int nM, int nmc, int ntotal, int niter);
+RcppExport SEXP _nlmixr2est_saemSeedLayoutTest_(SEXP nuSEXP, SEXP nu1BSEXP, SEXP nphi1SEXP, SEXP nphi0SEXP, SEXP nMixSEXP, SEXP nMSEXP, SEXP nmcSEXP, SEXP ntotalSEXP, SEXP niterSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type nu(nuSEXP);
+    Rcpp::traits::input_parameter< int >::type nu1B(nu1BSEXP);
+    Rcpp::traits::input_parameter< int >::type nphi1(nphi1SEXP);
+    Rcpp::traits::input_parameter< int >::type nphi0(nphi0SEXP);
+    Rcpp::traits::input_parameter< int >::type nMix(nMixSEXP);
+    Rcpp::traits::input_parameter< int >::type nM(nMSEXP);
+    Rcpp::traits::input_parameter< int >::type nmc(nmcSEXP);
+    Rcpp::traits::input_parameter< int >::type ntotal(ntotalSEXP);
+    Rcpp::traits::input_parameter< int >::type niter(niterSEXP);
+    rcpp_result_gen = Rcpp::wrap(saemSeedLayoutTest_(nu, nu1B, nphi1, nphi0, nMix, nM, nmc, ntotal, niter));
+    return rcpp_result_gen;
+END_RCPP
+}
 // saemEtaDistN_
 long saemEtaDistN_();
 RcppExport SEXP _nlmixr2est_saemEtaDistN_() {