+++ src/init.c
@@ -137,6 +137,7 @@ SEXP _nlmixr2est_nlmixr2Unscaled_(SEXP, SEXP);
 SEXP _nlmixr2est_saem_fit(SEXP);
 SEXP _nlmixr2est_saem_do_pred(SEXP, SEXP, SEXP);
 SEXP _nlmixr2est_saemPhi1RefineN_(void);
+SEXP _nlmixr2est_saemSeedLayoutTest_(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP _nlmixr2est_saemEtaDistN_(void);
 SEXP _nlmixr2est_saemEtaDistOn_(void);
 SEXP _nlmixr2est_rxEtaDistCorTest_(SEXP, SEXP);
@@ -395,6 +396,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"_nlmixr2est_saem_fit", (DL_FUNC) &_nlmixr2est_saem_fit, 1},
   {"_nlmixr2est_saem_do_pred", (DL_FUNC) &_nlmixr2est_saem_do_pred, 3},
   {"_nlmixr2est_saemPhi1RefineN_", (DL_FUNC) &_nlmixr2est_saemPhi1RefineN_, 0},
+  {"_nlmixr2est_saemSeedLayoutTest_", (DL_FUNC) &_nlmixr2est_saemSeedLayoutTest_, 9},
   {"_nlmixr2est_saemEtaDistN_", (DL_FUNC) &_nlmixr2est_saemEtaDistN_, 0},
   {"_nlmixr2est_saemEtaDistOn_", (DL_FUNC) &_nlmixr2est_saemEtaDistOn_, 0},
   {"_nlmixr2est_rxEtaDistCorTest_", (DL_FUNC) &_nlmixr2est_rxEtaDistCorTest_, 2},